src/video/generator.py
import cv2
import numpy as np
from typing import Dict, List, Tuple, Optional
import logging
from moviepy.editor import VideoFileClip, AudioFileClip, CompositeVideoClip, ColorClip
import tempfile
import os

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class VideoGenerator:
    def __init__(self, width: int = 1080, height: int = 1920):
        """Initialize the video generator with specified dimensions.
        
        Args:
            width (int): Width of the output video
            height (int): Height of the output video
        """
        self.width = width
        self.height = height
        self.fps = 30
        self.supported_themes = ['realistic', 'animated', 'abstract', 'cinematic']
        
    def generate_background(self, duration: float, theme: str,
                          audio_features: Dict[str, any]) -> str:
        """Generate a dynamic background video based on audio features.
        
        Args:
            duration (float): Duration of the video in seconds
            theme (str): Visual theme for the background
            audio_features (Dict): Dictionary containing audio analysis results
            
        Returns:
            str: Path to the generated video file
        """
        try:
            if theme.lower() not in self.supported_themes:
                raise ValueError(f"Unsupported theme. Supported themes: {self.supported_themes}")
            
            # Create base video clip
            base_clip = ColorClip((self.width, self.height), 
                                color=(0, 0, 0),
                                duration=duration)
            
            # Generate effects based on audio features
            effects = self._generate_effects(audio_features, theme)
            
            # Combine base clip with effects
            final_clip = CompositeVideoClip([base_clip] + effects)
            
            # Export video
            temp_file = tempfile.NamedTemporaryFile(delete=False, suffix='.mp4')
            final_clip.write_videofile(temp_file.name, 
                                     fps=self.fps,
                                     codec='libx264',
                                     audio=False)
            
            return temp_file.name
        except Exception as e:
            logger.error(f"Error generating background video: {str(e)}")
            raise
            
    def _generate_effects(self, audio_features: Dict[str, any], 
                         theme: str) -> List[VideoFileClip]:
        """Generate visual effects based on audio features and theme.
        
        Args:
            audio_features (Dict): Dictionary containing audio analysis results
            theme (str): Visual theme for the effects
            
        Returns:
            List[VideoFileClip]: List of effect clips
        """
        effects = []
        
        # Get sync points and energy profile
        sync_points = audio_features['sync_points']
        energy_profile = audio_features['energy_profile']
        
        if theme.lower() == 'abstract':
            effects.extend(self._generate_abstract_effects(sync_points, energy_profile))
        elif theme.lower() == 'realistic':
            effects.extend(self._generate_realistic_effects(sync_points, energy_profile))
        elif theme.lower() == 'animated':
            effects.extend(self._generate_animated_effects(sync_points, energy_profile))
        elif theme.lower() == 'cinematic':
            effects.extend(self._generate_cinematic_effects(sync_points, energy_profile))
            
        return effects
        
    def _generate_abstract_effects(self, sync_points: List[Tuple[float, float]],
                                 energy_profile: np.ndarray) -> List[VideoFileClip]:
        """Generate abstract visual effects (e.g., geometric patterns, particles).
        
        Args:
            sync_points (List[Tuple]): List of (time, strength) tuples
            energy_profile (np.ndarray): Audio energy profile
            
        Returns:
            List[VideoFileClip]: List of abstract effect clips
        """
        effects = []
        
        # Generate particle system
        particles = self._create_particle_system(sync_points, energy_profile)
        effects.append(particles)
        
        # Generate geometric patterns
        patterns = self._create_geometric_patterns(sync_points, energy_profile)
        effects.append(patterns)
        
        return effects
        
    def _generate_realistic_effects(self, sync_points: List[Tuple[float, float]],
                                  energy_profile: np.ndarray) -> List[VideoFileClip]:
        """Generate realistic visual effects (e.g., light flares, bokeh).
        
        Args:
            sync_points (List[Tuple]): List of (time, strength) tuples
            energy_profile (np.ndarray): Audio energy profile
            
        Returns:
            List[VideoFileClip]: List of realistic effect clips
        """
        effects = []
        
        # Generate light flares
        flares = self._create_light_flares(sync_points)
        effects.append(flares)
        
        # Generate bokeh effects
        bokeh = self._create_bokeh_effects(energy_profile)
        effects.append(bokeh)
        
        return effects
        
    def _create_particle_system(self, sync_points: List[Tuple[float, float]],
                              energy_profile: np.ndarray) -> VideoFileClip:
        """Create a particle system animation synchronized with the music.
        
        Args:
            sync_points (List[Tuple]): List of (time, strength) tuples
            energy_profile (np.ndarray): Audio energy profile
            
        Returns:
            VideoFileClip: Particle system animation
        """
        # Implementation for particle system generation
        # This is a placeholder - actual implementation would create
        # dynamic particle effects based on sync points and energy
        pass
        
    def _create_geometric_patterns(self, sync_points: List[Tuple[float, float]],
                                 energy_profile: np.ndarray) -> VideoFileClip:
        """Create geometric pattern animations synchronized with the music.
        
        Args:
            sync_points (List[Tuple]): List of (time, strength) tuples
            energy_profile (np.ndarray): Audio energy profile
            
        Returns:
            VideoFileClip: Geometric pattern animation
        """
        # Implementation for geometric pattern generation
        # This is a placeholder - actual implementation would create
        # dynamic geometric patterns based on sync points and energy
        pass
        
    def _create_light_flares(self, sync_points: List[Tuple[float, float]]) -> VideoFileClip:
        """Create light flare effects synchronized with strong beats.
        
        Args:
            sync_points (List[Tuple]): List of (time, strength) tuples
            
        Returns:
            VideoFileClip: Light flare effects
        """
        # Implementation for light flare generation
        # This is a placeholder - actual implementation would create
        # dynamic light flare effects based on sync points
        pass
        
    def _create_bokeh_effects(self, energy_profile: np.ndarray) -> VideoFileClip:
        """Create bokeh effects that respond to audio energy.
        
        Args:
            energy_profile (np.ndarray): Audio energy profile
            
        Returns:
            VideoFileClip: Bokeh effects
        """
        # Implementation for bokeh effect generation
        # This is a placeholder - actual implementation would create
        # dynamic bokeh effects based on energy profile
        pass
        
    def combine_with_audio(self, video_path: str, audio_path: str) -> str:
        """Combine video with audio track.
        
        Args:
            video_path (str): Path to the video file
            audio_path (str): Path to the audio file
            
        Returns:
            str: Path to the combined video file
        """
        try:
            # Load video and audio
            video = VideoFileClip(video_path)
            audio = AudioFileClip(audio_path)
            
            # Set audio
            final_video = video.set_audio(audio)
            
            # Export combined video
            temp_file = tempfile.NamedTemporaryFile(delete=False, suffix='.mp4')
            final_video.write_videofile(temp_file.name,
                                      fps=self.fps,
                                      codec='libx264')
            
            return temp_file.name
        except Exception as e:
            logger.error(f"Error combining video and audio: {str(e)}")
            raise
            
    def apply_effects(self, video_path: str, effects: List[Dict[str, any]]) -> str:
        """Apply post-processing effects to the video.
        
        Args:
            video_path (str): Path to the video file
            effects (List[Dict]): List of effect configurations
            
        Returns:
            str: Path to the processed video file
        """
        try:
            # Load video
            video = VideoFileClip(video_path)
            
            # Apply each effect
            for effect in effects:
                if effect['type'] == 'color_grading':
                    video = self._apply_color_grading(video, effect['params'])
                elif effect['type'] == 'blur':
                    video = self._apply_blur(video, effect['params'])
                # Add more effect types as needed
            
            # Export processed video
            temp_file = tempfile.NamedTemporaryFile(delete=False, suffix='.mp4')
            video.write_videofile(temp_file.name,
                                fps=self.fps,
                                codec='libx264')
            
            return temp_file.name
        except Exception as e:
            logger.error(f"Error applying effects: {str(e)}")
            raise
            
    def _apply_color_grading(self, video: VideoFileClip, 
                           params: Dict[str, float]) -> VideoFileClip:
        """Apply color grading to the video.
        
        Args:
            video (VideoFileClip): Input video
            params (Dict): Color grading parameters
            
        Returns:
            VideoFileClip: Color graded video
        """
        # Implementation for color grading
        # This is a placeholder - actual implementation would apply
        # color grading based on provided parameters
        pass
        
    def _apply_blur(self, video: VideoFileClip, 
                   params: Dict[str, float]) -> VideoFileClip:
        """Apply blur effect to the video.
        
        Args:
            video (VideoFileClip): Input video
            params (Dict): Blur parameters
            
        Returns:
            VideoFileClip: Blurred video
        """
        # Implementation for blur effect
        # This is a placeholder - actual implementation would apply
        # blur effect based on provided parameters
        pass 